+++ libs/serialization/test/src/unit_test_serialization_shared_ptr_polymorphic_version.cpp
@@ -67,7 +67,7 @@ public:
 
 	float GetValue() const override
 	{
-		return Base::GetValue() + b;
+		return Base::GetValue() + static_cast<float>(b);
 	}
 
 private:
