+++ libs/serialization/test/src/unit_test_serialization_weak_ptr_polymorphic.cpp
@@ -58,7 +58,7 @@ public:
 
 	float GetValue() const override
 	{
-		return Base::GetValue() + b;
+		return Base::GetValue() + static_cast<float>(b);
 	}
 
 private:
